minor PR feedback/style
